Add KRX gold spot (KRX 금현물) support in portfolio
- Scrape gold price from Naver Finance goldDailyQuote page
- Special asset code KRX_GOLD with auto-resolved name "KRX 금현물"
- Quantity accepts float (g unit) across save, edit, bulk import
- Frontend: search "금" to find KRX 금현물, display g unit for quantity,
  원/g tag, decimal input support
- Classified as Korean market in portfolio filter

+++ cache.py
@@ -822,7 +822,7 @@ async def get_portfolio(google_sub: str) -> list[dict]:
 
 
 async def save_portfolio_item(
-    google_sub: str, stock_code: str, stock_name: str, quantity: int, avg_price: float,
+    google_sub: str, stock_code: str, stock_name: str, quantity: float, avg_price: float,
     currency: str = "KRW",
 ) -> dict:
     db = await get_db()
